# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drachenboot Manager API",
        "version": "2.5.0",
        "description": "API documentation for the Drachenboot Manager application with multi-team support."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/events/{id}/canisters": {
            "post": {
                "summary": "Add Canister",
                "description": "Increments the canister count for an event.",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Event ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Canister added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "canisterCount": {
                                            "type": "integer"
                                        },
                                        "newCanisterId": {
                                            "type": "string",
                                            "description": "ID of the new canister (e.g., 'canister-1')"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to add canister",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Remove Canister",
                "description": "Decrements the canister count. Only works if there are unassigned canisters.",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Event ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "canisterId",
                        "in": "query",
                        "required": true,
                        "description": "Canister ID (e.g., 'canister-1')",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Canister removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Canister ID required or canister is assigned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to remove canister",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}/guests": {
            "post": {
                "summary": "Add Guest Paddler",
                "description": "Creates a guest paddler for a specific event with automatic attendance.",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Event ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "number"
                                    },
                                    "skills": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "drum",
                                                "steer"
                                            ]
                                        }
                                    }
                                },
                                "required": [
                                    "name",
                                    "weight",
                                    "skills"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest paddler created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaddlerFull"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to add guest",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}/guests/{guestId}": {
            "delete": {
                "summary": "Remove Guest Paddler",
                "description": "Deletes a guest paddler (cascade deletes attendance and assignments).",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Event ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "guestId",
                        "in": "path",
                        "required": true,
                        "description": "Guest paddler ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to remove guest",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaddlerFull": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique paddler identifier (CUID)"
                    },
                    "name": {
                        "type": "string"
                    },
                    "teamId": {
                        "type": "string",
                        "nullable": true,
                        "description": "Associated team ID"
                    },
                    "weight": {
                        "type": "number",
                        "description": "Weight in kg (redacted/0 for non-captains)"
                    },
                    "side": {
                        "type": "string",
                        "enum": [
                            "left",
                            "right",
                            "both"
                        ],
                        "nullable": true
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "drum",
                                "steer"
                            ]
                        }
                    },
                    "isGuest": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string",
                                "nullable": true
                            },
                            "email": {
                                "type": "string",
                                "nullable": true
                            },
                            "image": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "name",
                    "weight",
                    "skills",
                    "isGuest",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                },
                "required": [
                    "error"
                ]
            }
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "authjs.session-token",
                "description": "Session-based authentication using NextAuth.js. Requires active login session."
            }
        }
    },
    "security": [
        {
            "sessionAuth": []
        }
    ],
...
}